cas.o: Optimize encoding blocks (esp. when there are few edges)
Use VBR8 for encoding fixup and target info lists, and embed them in
BlockRef when they're small enough. The limit for embedding is
determined when writing and is controlled by
`--max-edges-to-embed-in-block`. The default is 8, which shaves off
15-20% total bytes for lib/Support/ objects within a single build. I'm
nervous to play around too much before we have numbers across builds,
since embedded lists won't get deduped. There's also still optimization
we could do here: the indexes in TargetInfoList are bigger than they
need to be.

Also inline unary target lists, depending on
`--inline-unary-target-lists`. This isn't having a big effect though,
and I noticed (again) that there is very little deduplication among
"target-lists". My suspicion is that the KeepAlive edges to the
`__eh_frame` are the root cause. I added a bunch of header documentation
to CompileUnitRef about a possible new schema.