cas.o: Simplify schema for abstract backedges
Stop constructing a special IndirectSymbolRef for abstract backedges.
Instead, use a reserved index in TargetInfo (equal to the number of
targets). Not using `-1ULL` here because that is expensive to encode in
VBR8.

This has almost no effect on its own, but allows a future commit to more
easily simplify IndirectSymbolRef.