Support: Allow SourceMgr to use a VFS
Add a SourceMgr constructor that takes a vfs::FileSystem, and a
mutator setFileSystem() for unittest convenience.

SourceMgr's only access to the filesystem is with
MemoryBuffer::getFile(). Rather than initializing SourceManager::FS with
vfs::getRealFileSystem(), allow it to be null and continue calling
MemoryBuffer::getFile() directly.

Maybe the move constructor should copy SourceMgr::FS, but it's not clear
that's valuable and it affects the refcount. For now this does the
default.