CAS.o: Generalize and optimzie data::TargetIndexList encoding
Generalize TargetIndexList encoding, to avoid requiring the Index to be
small (and drop comments suggesting optimizing based on this). In
NestedV1, every TargetIndex is from a specific edge-list, but in other
formats (such as FlatV1) there could be a bigger list of symbols being
indexed into.

Also pick up the optimization from FlatV1:

- Shift the index left to make room for a HasAddend bit.
- Only store the addend when it's non-zero.