[llvm-cas-object-format] Make FlatV1 schema ingestion deterministic for EH frames
LinkGraph creates an anonymous symbol that covers the full EH frame section
block and no edge points to it. The EH frame splitter may end up creating an
identical symbol (that has an edge pointed to it) which can result in
non-deterministic behavior in `FlatV1` schema ingestion, because the
symbols cannot be ordered in a deterministic manner.

Add a new function `removeRedundantEHFrameSymbol()` which removes this redundant
symbol before we run the EH frame splitter.