[DependencyScanning] Add ability to scan TU with multiple module imports
Update Dependency scanner so it can scan the dependency of a TU with
additional module imports that is not specified in the main TU.

This also unifies the DependencyScanningAction for TU and module. When
scanning for module dependency, it is equivalent to using an empty TU +
one addition module import.