[CAS] Add OnDiskHashMappedTrie
Add OnDiskHashMappedTrie. This is a on-disk persistent hash map that
uses a Trie data structure that is similar to TrieRawHashMap.
OnDiskHashMappedTrie is thread safe and process safe. It is mostly lock
free, except it internally coordinates cross process creation and
closing using file lock.

TODO: consider rename to OnDiskTrieHashMap to match TrieRawHashMap.