[DependencyScanning] Detect inconsistent file system during scanning
Teach DependencyScanning to find inconsistent file system during build
where the dependendent files are seen differently between different
modules and translation units.

This used to be detected when compiling and merging the CAS file system
but detecting it early in the dependendency scanning allows better error
message to be emitted and give more context to users for how to act when
receiving such an error.

As a result, the include-tree file system will no longer be nested and
is flatten out to save traveral time.

rdar://142914738